==> 1_preprocess_match_data.ipynb <==
import pandas as pd
from sqlalchemy import create_engine

database_path = r'D:\databases\cricket_data.db'
engine = create_engine(f'sqlite:///{database_path}')

# match_info

match_info = pd.read_sql_table('match_info', engine)
match_info

# matches

matches = pd.read_sql_table('matches', engine)
matches

# merge data

match_data = pd.merge(match_info, matches, on='id', how='outer')
match_data.rename(columns = {'id':'match_id'}, inplace = True)
match_data

# drop unneccesary columns

match_data = match_data.drop(columns=['series', 'season', 'link'])
match_data

# extract toss_winner and toss_decision from toss

def get_toss_winner(toss):
    if not isinstance(toss, str):
        return None

    if 'Northern Superchargers (Men)' in toss:
        return 'N S-Chargers'
    return toss.split(',')[0].strip()

def get_toss_decision(toss):
    if not isinstance(toss, str):
        return None
    toss = toss.lower()
    if 'bat' in toss:
        return 'bat'
    elif 'field' in toss:
        return 'field'
    else:
        return None

match_data['toss_winner'] = match_data['toss'].apply(get_toss_winner)
match_data['toss_decision'] = match_data['toss'].apply(get_toss_decision)
match_data = match_data.drop(columns=['toss'])
match_data

# extract match_date and format from match_days

def get_match_date(match_day):
    if not isinstance(match_day, str):
        return None
    return match_day.split('-')[0].strip()

def get_format(match_day):
    if not isinstance(match_day, str):
        return None
    format_str = match_day.split('(')[1].strip()
    overs = format_str.split('-')[0].strip()
    if overs in ['20', '50']:
        return overs
    else:
        return None

match_data['match_date'] = match_data['match_days'].apply(get_match_date)
match_data['format'] = match_data['match_days'].apply(get_format)
match_data = match_data.drop(columns=['match_days'])
match_data

# Convert match_dates to DD-MM-YYYY

def convert_date_format(date):
    if date is not None:
        date_obj = pd.to_datetime(date, format='%d %B %Y')
        return date_obj.strftime('%d-%m-%Y')
    else:
        return None

match_data['match_date'] = match_data['match_date'].apply(convert_date_format)
match_data

# extract match_winner from result

def get_match_winner(row):
    result = row['result']
    team1_name = row['team1_name']
    team2_name = row['team2_name']
    
    if not isinstance(result, str):
        return None

    if 'Match tied' in result:
        return 'Match tied'

    if 'No result' in result or 'abandoned' in result:
        return 'No result'
    
    if 'won' in result.lower():
        match_winner = result.split('won')[0].strip()
        if match_winner in team1_name:
            return team1_name
        elif match_winner in team2_name:
            return team2_name
        else:
            return match_winner
    else:
        return result

match_data['match_winner'] = match_data.apply(get_match_winner, axis=1)
match_data = match_data.drop(columns=['result'])
match_data

# handle match_winners

def handle_KnightRiders(row):
    match_winner = row['match_winner']
    team1_name = row['team1_name']
    team2_name = row['team2_name']
    
    if match_winner == 'KnightRiders':
        match_winner = 'Knight Riders'
        if match_winner in team1_name:
            return team1_name
        elif match_winner in team2_name:
            return team2_name
        else:
            return match_winner
    else:
        return match_winner

def handle_USA_and_UAE(match_winner):
    if match_winner=='U.S.A.':
        return 'United States of America'
    elif match_winner=='U.A.E.':
        return 'United Arab Emirates'
    else:
        return match_winner

def handle_short_team_names(match_winner):
    teams_dict = {'SF': 'San Francisco Unicorns', 'KKR': 'Kolkata Knight Riders', 'SRH': 'Sunrisers Hyderabad', 'RR': 'Rajasthan Royals',
              'RCB': 'Royal Challengers Bengaluru', 'LSG': 'Lucknow Super Giants', 'PBKS': 'Punjab Kings',
              'CSK': 'Chennai Super Kings', 'GT': 'Gujarat Titans', 'MI': 'Mumbai Indians'}
    
    if match_winner in teams_dict.keys():
        return teams_dict[match_winner]
    else:
        return match_winner

def handle_DC(row):
    match_winner = row['match_winner']
    team1_name = row['team1_name']
    team2_name = row['team2_name']

    if match_winner == 'DC':
        if 'Dubai Capitals' in [team1_name, team2_name]:
            return 'Dubai Capitals'
        elif 'Delhi Capitals' in [team1_name, team2_name]:
            return 'Delhi Capitals'
        else:
            return match_winner
    else:
        return match_winner

match_data['match_winner'] = match_data.apply(handle_KnightRiders, axis=1)
match_data['match_winner'] = match_data['match_winner'].apply(handle_USA_and_UAE)
match_data['match_winner'] = match_data['match_winner'].apply(handle_short_team_names)
match_data['match_winner'] = match_data.apply(handle_DC, axis=1)
match_data

def valid_match_winner(row):
    match_winner = row['match_winner']
    team1_name = row['team1_name']
    team2_name = row['team2_name']
    if match_winner in [team1_name, team2_name]:
        return True
    else:
        return False

def show_invalid_match_winners(data):
    df = data.copy()
    df['valid_match_winner'] = df.apply(valid_match_winner, axis=1)
    return df[df['valid_match_winner']==False]

df = show_invalid_match_winners(match_data)
df['match_winner'].unique()

# handle team_scores

def handle_team_score(team_score):
    if '/' in team_score:
        return team_score.split('/')[0].strip()
    elif team_score=='':
        return None
    else:
        return team_score

match_data['team1_score'] = match_data['team1_score'].apply(handle_team_score)
match_data['team2_score'] = match_data['team2_score'].apply(handle_team_score)
match_data

# handle nan values

match_data = match_data.dropna(subset=['ground'])
match_data

# VERIFY

# info

match_data.info()

# nan values

match_data.isnull().sum()

## toss_winner

invalid_toss_winners = match_data[~match_data['toss_winner'].isin(match_data[['team1_name', 'team2_name']].values.ravel())]['toss_winner'].unique()
invalid_toss_winners

## toss_decision

match_data['toss_decision'].unique()

## match_date

def check_date_format(date):
    try:
        pd.to_datetime(date, format='%d-%m-%Y')
        return False
    except ValueError:
        return True

incorrect_dates_df = match_data[match_data['match_date'].apply(check_date_format)]
incorrect_dates_df

## format

match_data['format'].unique()

## match_winner

invalid_winner_rows = match_data[~match_data['match_winner'].isin(match_data[['team1_name', 'team2_name']].values.ravel())]['match_winner'].unique()
invalid_winner_rows

## team1_score

match_data['team1_score'].astype(float)

## team2_score

match_data['team2_score'].astype(float)

# Final State

match_data

match_data.to_csv('data/match_data.csv', index=False)
