==> 2_preprocess_performance_data.ipynb <==
import pandas as pd
from sqlalchemy import create_engine

database_path = r'D:\databases\cricket_data.db'
engine = create_engine(f'sqlite:///{database_path}')

# batting_data

batting_data = pd.read_sql_table('batting_data', engine)
batting_data

# drop unneccesary columns from batting_data

batting_data = batting_data.drop(columns=['dismissal', 'balls', 'maidens', 'fours', 'sixes', 'strike_rate'])
batting_data

# clean player names in batting_data

import re

def clean_player_name(name):
    name = re.sub(r'\(.*?\)', '', name)
    name = re.sub(r'[^a-zA-Z0-9\s]', '', name)
    return name.strip()


batting_data['player'] = batting_data['player'].apply(clean_player_name)
batting_data

# bowling_data

bowling_data = pd.read_sql_table('bowling_data', engine)
bowling_data

# drop unneccesary columns from bowling_data

bowling_data = bowling_data.drop(columns=['id', 'overs', 'maidens', 'runs', 'economy', 'dots', 'fours',  'sixes', 'wides', 'no_balls'])
bowling_data

# clean player names in bowling_data

bowling_data['player'] = bowling_data['player'].apply(clean_player_name)
bowling_data

# merged_data

performance_data = pd.merge(batting_data, bowling_data, on=['player', 'match_id', 'team_name'], how='outer')
performance_data = performance_data.sort_values(by=['match_id', 'id']).reset_index(drop=True)
performance_data['performance_id'] = range(1, len(performance_data) + 1)
performance_data = performance_data.drop(columns=['id'])
performance_data

# rearrange columns

performance_data = performance_data[['performance_id', 'player', 'match_id', 'team_name', 'runs', 'wickets']]
performance_data

# info and null values

performance_data.info()

performance_data.isnull().sum()

# Final State

performance_data

# Save

performance_data.to_csv('data/performance_data.csv', index=False)
