==> 3_preprocessing_final.ipynb <==
import pandas as pd

match_data = pd.read_csv('data/match_data.csv')
match_data

performance_data = pd.read_csv('data/performance_data.csv')
performance_data

# merge data

data = pd.merge(performance_data, match_data, on='match_id', how='outer')
data

# clean team_name

def handle_incomplete_names(row):
    team_name = row['team_name']
    team1_name = row['team1_name']
    team2_name = row['team2_name']

    if team_name == 'KnightRiders':
        team_name = 'Knight Riders'
    
    if team_name in team1_name:
        return team1_name
    elif team_name in team2_name:
        return team2_name
    else:
        return team_name

def handle_short_team_names(team_name):
    teams_dict = {'SF': 'San Francisco Unicorns', 'KKR': 'Kolkata Knight Riders', 'SRH': 'Sunrisers Hyderabad', 'RR': 'Rajasthan Royals',
              'RCB': 'Royal Challengers Bengaluru', 'LSG': 'Lucknow Super Giants', 'PBKS': 'Punjab Kings',
              'CSK': 'Chennai Super Kings', 'GT': 'Gujarat Titans', 'MI': 'Mumbai Indians', 'P.N.G.': 'Papua New Guinea',
              'U.S.A.': 'United States of America', 'U.A.E.': 'United Arab Emirates'}
    
    if team_name in teams_dict.keys():
        return teams_dict[team_name]
    else:
        return team_name

def handle_DC(row):
    team_name = row['team_name']
    team1_name = row['team1_name']
    team2_name = row['team2_name']

    if team_name == 'DC':
        if 'Dubai Capitals' in [team1_name, team2_name]:
            return 'Dubai Capitals'
        elif 'Delhi Capitals' in [team1_name, team2_name]:
            return 'Delhi Capitals'
        else:
            return team_name
    else:
        return team_name




data['team_name'] = data.apply(handle_incomplete_names, axis=1)
data['team_name'] = data['team_name'].apply(handle_short_team_names)
data['team_name'] = data.apply(handle_DC, axis=1)
data

def valid_team_name(row):
    team_name = row['team_name']
    team1_name = row['team1_name']
    team2_name = row['team2_name']
    if team_name in [team1_name, team2_name]:
        return True
    else:
        return False

def show_invalid_team_names(data):
    df = data.copy()
    df['valid_team_name'] = df.apply(valid_team_name, axis=1)
    return df[df['valid_team_name']==False]

df = show_invalid_team_names(data)
df

# handle null values

data['runs'] = data['runs'].fillna(0)
data['wickets'] = data['wickets'].fillna(0)
data = data.dropna(subset=['format'])
data

# change data_types

data.loc[:, 'wickets'] = data['wickets'].astype(int)
data.loc[:, 'format'] = data['format'].astype(int)

data.loc[:, 'runs'] = pd.to_numeric(data['runs'], errors='coerce')
data.loc[:, 'runs'] = data['runs'].fillna(0)

data.loc[:, 'runs'] = data['runs'].astype(int)

# Verify

data.isnull().sum()

data.info()

# Final State

data

data.to_csv('data/final_preprocessed_dataset.csv', index=False)
