==> 5_model.ipynb <==
import pandas as pd

data = pd.read_csv('data/final_data.csv')
data['match_date'] = pd.to_datetime(data['match_date'], format="%Y-%m-%d")
data

# split_data

from datetime import datetime, timedelta

split_date = '2024-10-8'
split_date_obj = datetime.strptime(split_date, '%Y-%m-%d')

prediction_date_obj = split_date_obj + timedelta(days=1)
prediction_date = prediction_date_obj.strftime('%Y-%m-%d')

print("Split Date:", split_date)
print("Prediction Date:", prediction_date)


train_data = data[data['match_date'] < split_date]
test_data = data[data['match_date'] >= split_date]

X_train = train_data.drop(columns=['player', 'performance_id', 'match_date', 'fantasy_score', 'top_performer'])
y_train = train_data['top_performer']

X_test = test_data.drop(columns=['player', 'performance_id', 'match_date', 'fantasy_score', 'top_performer'])
y_test = test_data['top_performer']

# model

from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.svm import SVC
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import GaussianNB

# model = RandomForestClassifier(n_estimators=100, random_state=42)
model = LogisticRegression()
# model = DecisionTreeClassifier(random_state=104)
# model = SVC(kernel='rbf')
# model = KNeighborsClassifier(n_neighbors=5)
# model = GaussianNB()

model.fit(X_train, y_train)

y_pred = model.predict(X_test)

from sklearn.metrics import r2_score, mean_absolute_error, mean_squared_error, accuracy_score

# print(f'r2_score = {r2_score(y_pred, y_test)}')
# print(f'mean_absolute_error = {mean_absolute_error(y_pred, y_test)}')
# print(f'mean_squared_error = {mean_squared_error(y_pred, y_test)}')
ac = accuracy_score(y_pred, y_test)
print(f'accuracy_score = {ac}')
print(f'team_accuracy = {ac**(22)*100}%')
print(f'correct team prediction chance: 1 in {ac**(-22)} matches')

# prediction

data[data['match_date']>=prediction_date]['match_id'].unique()

def predict_match(match_id, data):
    match_data = data[data['match_id'] == match_id]
    X_prediction = match_data.drop(columns=['player', 'performance_id', 'match_date', 'fantasy_score', 'top_performer'])
    y_prediction = model.predict(X_prediction)
    
    prediction_data = match_data[['player', 'top_performer']]
    prediction_data['prediction'] = y_prediction
    
    return prediction_data

prediction_data = predict_match(1, data)
prediction_data

# prediction_1 important_but_not_important

def find_rank(player, rank_list):
    return (rank_list.index(player)) + 1

def predict_match(match_id, data):
    prediction_data = data[data['match_id'] == match_id]
    X_prediction = prediction_data.drop(columns=['player', 'performance_id', 'match_date', 'fantasy_score'])
    y_prediction = model.predict(X_prediction)
    
    prediction_data = prediction_data[['match_id', 'player', 'fantasy_score']]
    prediction_data['predicted_fantasy_score'] = y_prediction

    predicted_rank_list = prediction_data.sort_values(by=['predicted_fantasy_score'])['player'].tolist()
    predicted_rank_list.reverse()

    actual_rank_list = prediction_data.sort_values(by=['fantasy_score'])['player'].tolist()
    actual_rank_list.reverse()

    prediction_data['actual_rank'] = prediction_data['player'].apply(lambda x: find_rank(x, actual_rank_list))
    prediction_data['predicted_rank'] = prediction_data['player'].apply(lambda x: find_rank(x, predicted_rank_list))

    
    return prediction_data

def predict_matches(start_date, end_date, data):
    matches_data = data[data['match_date'].between(start_date, end_date)]
    match_ids = matches_data['match_id'].unique().tolist()

    prediction_data = pd.DataFrame({})
    for match_id in match_ids:
        prediction_data = pd.concat([prediction_data, predict_match(match_id, data)], ignore_index=True)
    return prediction_data

start_date = prediction_date
end_date = '2024-10-04'

prediction1_data = predict_matches(start_date, end_date, data)
prediction1_data

prediction1_data[prediction1_data['match_id']==5]

# prediction important_but_not_important

def predict_match(match_id, data):
    prediction_data = data[data['match_id'] == match_id]
    X_prediction = prediction_data.drop(columns=['player', 'performance_id', 'match_date', 'fantasy_score'])
    y_prediction = model.predict(X_prediction)
    prediction_data = prediction_data[['player', 'fantasy_score']]
    prediction_data['predicted_fantasy_score'] = y_prediction
    
    predicted_sorted_df = prediction_data.sort_values(by=['predicted_fantasy_score']).iloc[-11:]
    dream_sorted_df = prediction_data.sort_values(by=['fantasy_score']).iloc[-11:]
    
    predicted_team_score = prediction_data.sort_values(by=['predicted_fantasy_score']).iloc[-11:]['fantasy_score'].sum()
    dream_team_score = prediction_data.sort_values(by=['fantasy_score']).iloc[-11:]['fantasy_score'].sum()
    
    predicted_XI = prediction_data.sort_values(by=['predicted_fantasy_score']).iloc[-11:]['player'].tolist()
    predicted_XI.sort()
    
    dream_XI = prediction_data.sort_values(by=['fantasy_score']).iloc[-11:]['player'].tolist()
    dream_XI.sort()
    
    common_players = list(filter(lambda x: x in predicted_XI, dream_XI))
    
    return predicted_sorted_df, dream_sorted_df, predicted_team_score, dream_team_score, common_players

data[data['match_date'] == prediction_date]['match_id'].unique()

predicted_sorted_df, dream_sorted_df, predicted_team_score, dream_team_score, common_players = predict_match(8, data)

predicted_team_score

dream_team_score

print(len(common_players))
common_players

predicted_sorted_df

dream_sorted_df
